+++ Tools/WarningsPool.cs
@@ -37,7 +37,11 @@ namespace MyBox
 
 		public static bool LogWarning(UnityEditor.SerializedProperty property, string message, Object target = null)
 			=> LogWarning($"Property <color=brown>{property.name}</color> " +
-			              $"in object <color=brown>{property.serializedObject.targetObject.name}</color> caused: " + message, target);
+			              $"in Object <color=brown>{property.serializedObject.targetObject.name}</color> caused: " + message, target);
+
+		public static bool LogCollectionsNotSupportedWarning(UnityEditor.SerializedProperty property, string nameOfType)
+			=> LogWarning(property, $"Array fields are not supported by <color=brown>[{nameOfType}]</color>. " +
+			                        "Consider to use <color=blue>CollectionWrapper</color>", property.serializedObject.targetObject);
 		#endif